username name accessbility added!!

+++ src/app/api/route.ts
@@ -1,14 +1,17 @@
-import { mongoDBConnection } from "@/dbConfig/DbConfig";
-import { FolderFileModel } from "@/models/FolderFiles";
 import { NextApiRequest } from "next";
+import { mongoDBConnection } from "@/dbConfig/DbConfig";
+import { FolderFileModel, IFile } from "@/models/FolderFiles";
+import { NextRequest } from "next/server";
+import { UserModel } from "@/models/UserModel";
 
 mongoDBConnection();
-export async function GET(req: Request, res: Response) {
+export async function GET(req: NextRequest, res: Response) {
   try {
-    const files = await FolderFileModel.find({});
+    const userName = req.cookies.get("token")?.value;
+    const user = await UserModel.findOne({ userName }).populate("folders");
 
-    if (files) {
-      return Response.json(files, {
+    if (user) {
+      return Response.json(user.folders, {
         status: 200,
       });
     }
@@ -24,25 +27,77 @@ export async function GET(req: Request, res: Response) {
     console.log(error);
   }
 }
-export async function POST(req: Request) {
+export async function POST(req: NextRequest) {
   try {
+    const files: IFile[] = [
+      {
+        fileName: "index",
+        fileType: "html",
+        value: `
+      <!DOCTYPE html>
+<html lang="en">
+<head>
+      <meta charset="UTF-8">
+      <meta name="viewport" content="width=device-width, initial-scale=1.0">
+      <title>Document</title>
+</head>
+<body>
+      <h1>Hello World!</h1>
+</body>
+</html>
+      `,
+      },
+      {
+        fileName: "index",
+        fileType: "css",
+        value: `
+        * {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+        `,
+      },
+      {
+        fileName: "index",
+        fileType: "js",
+        value: `
+        console.log("Hello World");
+        `,
+      },
+    ];
     const { folderName } = await req.json();
-    if (folderName) {
-      const newFolder = new FolderFileModel({
-        folderName,
-      });
-
-      const savedFolder = await newFolder.save();
-      if (savedFolder) {
-        return Response.json({
-          message: "New Folder is added",
-          data: savedFolder,
+    const userName = req.cookies.get("token")?.value;
+    const user = await UserModel.findOne({
+      userName,
+    });
+    if (user) {
+      if (folderName) {
+        const newFolder = new FolderFileModel({
+          folderName,
+          files,
         });
+        user.folders.push(newFolder._id);
+        await user.save();
+        const savedFolder = await newFolder.save();
+        if (savedFolder) {
+          return Response.json({
+            message: "New Folder is added",
+            data: savedFolder,
+          });
+        }
+      } else {
+        return Response.json(
+          {
+            message: "Some issues occured!",
+          },
+          { status: 404 }
+        );
       }
     } else {
       return Response.json(
         {
-          message: "Some issues occured!",
+          message: "No user is Found!!",
         },
         { status: 404 }
       );
@@ -51,9 +106,12 @@ export async function POST(req: Request) {
     console.log(error);
   }
 }
-export async function PATCH(req: Request) {
+export async function PATCH(req: NextRequest) {
+  const userName = req.cookies.get("token")?.value;
+
   const { id } = await req.json();
   console.log(id);
+
   try {
     if (!id) {
       Response.json(
@@ -64,10 +122,43 @@ export async function PATCH(req: Request) {
       );
       return;
     }
-    const deleteFolder = await FolderFileModel.findByIdAndDelete(id);
+    if (userName) {
+      const user = await UserModel.findOneAndUpdate(
+        { userName },
+        {
+          $pull: { folders: id },
+        }
+      );
+      const deleteFolder = await FolderFileModel.findByIdAndDelete(id);
+      return Response.json(
+        {
+          message: "Folder is Deleted Successfully!",
+        },
+        {
+          status: 200,
+        }
+      );
+    }
+  } catch (error) {
+    console.log(error);
+  }
+}
+
+export async function PUT(req: NextRequest) {
+  try {
+    const { folderId, _id, value } = await req.json();
+
+    const folder = await FolderFileModel.findOneAndUpdate(
+      { _id: folderId, "files._id": _id },
+      {
+        $set: {
+          "files.$.value": value,
+        },
+      }
+    );
     return Response.json(
       {
-        message: "Folder is Deleted Successfully!",
+        message: "Updated code Successfully!",
       },
       {
         status: 200,
